chore: move log message matchers up
